Bound the chooser to the square picker
The color chooser (white circle) could be dragged out the bounds of the
square picker. The code is refactored so the cirlce couldn't be dragged
out of the bounds of the picker. The xy_to_sv function had most of it's
logic removed since it's all handled in gesture_drag_update.